Fix egress firewall tests by updating the URL from docs.openshift.com to redhat.com after docs migration
docs.openshift.com got moved to redhat.com and our egress firewall/egressnetworkpolicy tests were
using that fqdn to curl and test for egress firewall functionality. With the DNS name no longer resolving,
these tests need to be updated to the redhat.com server instead. More importantly, redhat.com has a longer
TTL for DNS entries and that reduces the surface area for the pod and nodes to get different IPs for the
FQDN.